+++ src/main/java/com/example/catshopbackend/repositories/ProductRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 
 import java.util.List;
 
-public interface ProductRepository extends MongoRepository<Product, Integer> {
+public interface ProductRepository extends MongoRepository<Product, Long> {
 
     List<Product> findByCategory(String category);
 }
